chore: set Symfony compatibility to 5.4, 6.4 and 7.0
