--- src/Knp/DictionaryBundle/Symfony/SupportedVersion.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Knp\DictionaryBundle\Symfony;
-
-final class SupportedVersion
-{
-    public const v5_4 = '5.4';
-    public const v6_2 = '6.2';
-    public const v6_3 = '6.3';
-
-    private function __construct() {}
-}
